requirements_unfixed.txt
google-cloud-bigquery==2.3.1
jupyterlab==2.2.9
pytest
pytest-ordering
black
coverage
sphinx
sphinx-rtd-theme
flake8
setuptools
pandas
matplotlib
numpy
cython
lxml
scipy
eppy
plotly
ipywidgets
psutil
attrs
google-cloud-storage
pandas-gbq
tqdm
numba
gcsfs
dask
requests
scikit-learn
statsmodels
tzwhere
numexpr
bottleneck
casadi
